# Optimize logging for journald